refactor: improved  PetDocumentsView
