+++ react-client/src/components/Pet/PetDocumentsView.tsx
@@ -1,9 +1,11 @@
 import { Meta, UppyFile } from "@uppy/core";
-import { useRef, useState } from "react";
-import { Button, Card, DragAndDropZone, Icon, Text } from "../design-system";
+import { useState } from "react";
+import { Text } from "../design-system";
 import { PetCardRecord } from "./PetCardRecord";
 import { PetRecord } from "./types/PetRecordsTypes";
 import { useUppyUploader } from "~/hooks/useUppyUploader";
+import { PetDocumentsDragAndDrop } from "./PetDocumentsDragAndDrop";
+import { simplifyFileType } from "~/util/fileUtil";
 
 type PetDocumentViewProps = {
   documents: PetRecord[];
@@ -16,7 +18,6 @@ export const PetDocumentsView = ({
   onDelete,
   recordType,
 }: PetDocumentViewProps) => {
-  const fileInputRef = useRef<HTMLInputElement>(null);
   const [file, setFile] = useState<UppyFile<Meta, Record<string, never>>>();
 
   const { isUploading, handleFiles } = useUppyUploader({
@@ -45,62 +46,17 @@ export const PetDocumentsView = ({
         Upload and attach files
       </Text>
 
-      <DragAndDropZone
-        handleFiles={handleFiles}
-        className="rounded-2xl border-solid border-brand-main"
-      >
-        <input
-          id="drop-input"
-          ref={fileInputRef}
-          className="sr-only"
-          type="file"
-          onChange={handleInputChange}
-        />
-        <Card>
-          <Button
-            aria-label="Upload document"
-            className="w-full rounded-none"
-            onClick={handleUploadClick}
-            variant="link"
-          >
-            <div className="grid place-items-center px-large py-base">
-              <div className="pb-small">
-                <Icon
-                  display="uploadCloud"
-                  className="text-brand-main"
-                  size={32}
-                />
-              </div>
-              <Text fontFamily="raleway" fontWeight="bold" size="sm">
-                Click to upload or drag and drop
-              </Text>
-              <Text color="tertiary" size="xs">
-                PNG, JPG, PDF, TXT, DOC, DOCX (max 10Mb)
-              </Text>
-            </div>
-          </Button>
-        </Card>
-      </DragAndDropZone>
+      <PetDocumentsDragAndDrop handleFiles={handleFiles} />
 
-      {isUploading && typeof file?.name !== "undefined" && (
-        // TODO: 81832 fix this after implementing all logic to upload documents
+      {isUploading && file?.name && (
         <PetCardRecord
-          record={{ fileName: file.name, fileType: "pdf" }}
+          record={{
+            fileName: file.name,
+            fileType: simplifyFileType(file.type),
+          }}
           isUploadingFile={isUploading}
         />
       )}
-
-      <div id="uppy-dashboard"></div>
     </div>
   );
-
-  function handleUploadClick() {
-    fileInputRef.current?.click();
-  }
-
-  function handleInputChange(event: React.ChangeEvent<HTMLInputElement>) {
-    if (event.target.files) {
-      handleFiles(event.target.files);
-    }
-  }
 };
